+++ migrations/migrations/004-users-extend.js
@@ -1,0 +1,7 @@
+const query = `
+  ALTER TABLE users
+    ADD COLUMN admin boolean DEFAULT false,
+    ADD COLUMN organisation UUID REFERENCES organisations (id);
+`;
+
+module.exports.generateSql = () => `${query}`;