+++ pages/api/auth/[...nextauth].js
@@ -1,5 +1,8 @@
 import NextAuth from "next-auth";
 import Providers from "next-auth/providers";
+import Adapters from "next-auth/adapters";
+
+import Models from "../../../migrations/models";
 
 export default NextAuth({
   // Configure one or more authentication providers
@@ -13,4 +16,12 @@ export default NextAuth({
 
   // A database is optional, but required to persist accounts in a database
   database: process.env.DATABASE_URL ?? null,
+
+  adapter: process.env.DATABASE_URL
+    ? Adapters.TypeORM.Adapter(process.env.DATABASE_URL, {
+        models: {
+          User: Models.User,
+        },
+      })
+    : null,
 });