+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Use a cache to fill frequent requests to the Template API to reduce overall load on Lambdas
+- Enable support for development using local Lambdas
+- Enable support for organisations and organisation management through the admin panel
 
 ### Changed